波次详情查询修改

+++ oms/src/main/java/com/upedge/oms/modules/order/entity/Order.java
@@ -136,7 +136,7 @@ public class Order{
 
 	private Integer packState;
 
-	private Integer packNo;
+	private Long packNo;
 
 	public Order() {
 	}
